bootloader: Copy over code from a previous bootloader project.
The bootloader doesn't currently do anything except for the following:

* Initializes a logger
* Initializes a really simple allocator
* Disables the watchdog timer
* Log firmware information
* Open ESP file system (doesn't actually do anything with the files yet)

A function for opening a file from a path is also included, but not yet
used (`open_file`). Before opening the file, it walks the path and tries
to open each parent directory of the file, starting with the root.
Returns the file handle if it was opened successfully.