bootloader: Update `FileDescriptor` to include a custom data type
This allows file systems to store metadata and APIs that are unique to
a single file system.

Some advanced trait trickery is needed for this, which makes a lot of
the code excessively verbose. Also, I'm not sure how well this solution
works in general. :P

Time will tell whether I revert this change or not...