bootloader: Got a proof of concept file system interface working
This implementation mostly uses raw pointers, so is inherently unsafe. I
plan to thoroughly review the interface so that I might be able to change
the file system functions to not be marked unsafe.

Also, I plan to change the interface to only use raw pointers internally.
This can prevent callers of the interface functions from messing things
up, but time will tell if I can implement it well.