aarch64/entry: Got serial output working in assembler
For some reason, the Rust code doesn't properly output to the UART,
while the assembly code does. I wonder if it could be something to do
with improper linking or relocation? It seems that one of the branch
instructions might be going to the wrong address and not running the
Rust UART code.