aarch64: Tried to use intrusive lists to manage free memory
While this is a good idea, I will probably never use this commit. This
is an extremely rough implementation that works as a proof of concept.
My next task is to reimplement intrusive lists so they are easy to use.

I am keeping this commit so I can use it as a reference.