slab_allocator: Working on an algorithm to allocate multiple slabs
This has been too difficult to deal with because of the many indexing
calculations needed when working with bitmaps. It would be much easier
if I could simply use a Rust `Iterator` to iterate over each bit in
order. This way, the slab allocator does not need to manually calculate
bit indices, mask bits, or deal with byte boundaries.

Thus, I'll be working on a bitmap iterator next so that I can complete
this algorithm.