This commit is a mess of stuff I've been working on for a while
Mostly related to creating a physical memory allocator, but I also
removed a bunch of deprecated stuff that I plan on fully reimplementing.

Also removed the book, as I'm working on it in a separate location for
now. It will be put back into a repository once I get more organized.

As for the physical allocator, it does not work, and has not even been
tested yet. I'm still working on the details of merging and splitting
memory. But, I felt that I should do a new commit to save my progress so
far, because my allocator design seems to be decent. Time will tell if
it actually works... :L